sponge: use typestate pattern
